+++ .eslintignore
@@ -1,7 +1,5 @@
 App.js
-Home.jsx
 
-Button.jsx
 index.js
 
 Utils.js
